implemented O365 user token based authentication
