+++ DocumentService.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DocumentService", "DocumentService\DocumentService.csproj", "{655B6294-AF6F-4DAC-BFFD-80B0222B0030}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Net4UserTokenLib", "Net4UserTokenLib\Net4UserTokenLib.csproj", "{35DD060B-7224-4762-BD5E-5E73063A4914}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{655B6294-AF6F-4DAC-BFFD-80B0222B0030}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{655B6294-AF6F-4DAC-BFFD-80B0222B0030}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{655B6294-AF6F-4DAC-BFFD-80B0222B0030}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35DD060B-7224-4762-BD5E-5E73063A4914}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35DD060B-7224-4762-BD5E-5E73063A4914}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35DD060B-7224-4762-BD5E-5E73063A4914}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35DD060B-7224-4762-BD5E-5E73063A4914}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
