+++ DocumentService/Global.asax.cs
@@ -7,7 +7,7 @@ using System.Web.Mvc;
 using System.Web.Optimization;
 using System.Web.Routing;
 
-namespace AttachmentExampleService
+namespace AttachmentService
 {
   // Note: For instructions on enabling IIS6 or IIS7 classic mode, 
   // visit http://go.microsoft.com/?LinkId=9394801
@@ -16,12 +16,7 @@ namespace AttachmentExampleService
   {
     protected void Application_Start()
     {
-      AreaRegistration.RegisterAllAreas();
-
-      WebApiConfig.Register(GlobalConfiguration.Configuration);
-      FilterConfig.RegisterGlobalFilters(GlobalFilters.Filters);
-      RouteConfig.RegisterRoutes(RouteTable.Routes);
-      BundleConfig.RegisterBundles(BundleTable.Bundles);
+        GlobalConfiguration.Configure(WebApiConfig.Register);
     }
   }
 }
