+++ DocumentService/Models/UserIdRequest.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace AttachmentService.Models
+{
+    public class UserIdRequest
+    {
+        public string token;
+    }
+}
