+++ Net4UserTokenLib/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.Exchange.WebServices" version="2.2" targetFramework="net40" />
+  <package id="Microsoft.IdentityModel" version="6.1.7600.16394" targetFramework="net40" />
+</packages>
