+++ README.md
@@ -1,1 +1,5 @@
 # ibera-document-service
+
+Before running the solution, ensure you set `StorageAccountConnectionString` and     `OutlookIntegrationHostUri` (e.g. https://localhost:8443/Functions.html)
+
+When running local you can set it in your [web.config](https://github.com/CatalystCode/ibera-document-service/blob/master/DocumentService/Web.config) or in AppSettings when deploying to an Azure WebApp.
